Initial work to redefine the F# unit type as an empty struct.
The current code compiles using the lkg compiler (or a proto compiler
built without these changes). However, after compiling in these changes,
the compiler is no longer able to bootstrap itself because some phase of
the compiler expects unit to be a reference type that can simply be
replaced with a null value; this results in the compiler crashing or
succeeding but generating invalid IL (which then crashes when run).